Move parser tests to their own file
